+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This file is used to track changes made to the project over time.
 #### Spatial
 - Added `action_matrix`, `dual_matrix`, and `inv_matrix` methods to SE3 transforms to compute the action, dual action, and inverse action matrices.
 - Add more bindings for inertias.
+- Better display for spatial types.
 
 ### Fixed
 - Better handling of infinite configurations for joint subspaces.